pyproject.toml
[project]
name = "intertia-django-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "inertia-django>=1.2.0",
]
